6-prefect/pyproject.toml
[project]
name = "is-566-08-prefect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "pandas>=3.0.1",
    "prefect>=3,<4",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
